Nav adjusted
Only links to the right are avaliable to all. The links to the right, will change depending on the login. When the user is loged in, will appear a sandwiche menu and the name of the user. In the sandwiche menu, options like: destinations, dashboard, my infromation, etc (all personal data).